+++ pkg/devfile/adapters/common/utils.go
@@ -1,30 +1,61 @@
 package common
 
 import (
-	"fmt"
-	"reflect"
+	"os"
 
 	"github.com/golang/glog"
-	"github.com/pkg/errors"
 
 	"github.com/openshift/odo/pkg/devfile/versions"
 	"github.com/openshift/odo/pkg/devfile/versions/common"
 )
 
+// PredefinedDevfileCommands encapsulates constants for predefined devfile commands
 type PredefinedDevfileCommands string
 
 const (
+	// DefaultDevfileBuildCommand is a predefined devfile command for build
 	DefaultDevfileBuildCommand PredefinedDevfileCommands = "devbuild"
-	DefaultDevfileRunCommand   PredefinedDevfileCommands = "devrun"
+
+	// DefaultDevfileRunCommand is a predefined devfile command for run
+	DefaultDevfileRunCommand PredefinedDevfileCommands = "devrun"
+
+	// SupervisordInitContainerName The init container name for supervisord
+	SupervisordInitContainerName = "copy-supervisord"
+
+	// Default Image that will be used containing the supervisord binary and assembly scripts
+	// use GetBootstrapperImage() function instead of this variable
+	defaultBootstrapperImage = "registry.access.redhat.com/openshiftdo/odo-init-image-rhel7:1.0.2"
+
+	// SupervisordVolumeName Create a custom name and (hope) that users don't use the *exact* same name in their deployment (occlient.go)
+	SupervisordVolumeName = "odo-supervisord-shared-data"
+
+	// SupervisordMountPath The supervisord Mount Path for the container mounting the supervisord volume
+	SupervisordMountPath = "/opt/odo/"
+
+	// SupervisordBinaryPath The supervisord binary path inside the container volume mount
+	SupervisordBinaryPath = "/opt/odo/bin/supervisord"
+
+	// SupervisordConfFile The supervisord configuration file inside the container volume mount
+	SupervisordConfFile = "/opt/odo/conf/devfile-supervisor.conf"
+
+	// OdoInitImageContents The path to the odo init image contents
+	OdoInitImageContents = "/opt/odo-init/."
+
+	// ENV variable to overwrite image used to bootstrap SupervisorD in S2I and Devfile builder Image
+	bootstrapperImageEnvName = "ODO_BOOTSTRAPPER_IMAGE"
 )
 
 func isComponentSupported(component common.DevfileComponent) bool {
 	// Currently odo only uses devfile components of type dockerimage, since most of the Che registry devfiles use it
-	if component.Type != common.DevfileComponentTypeDockerimage {
-		return false
-	}
+	return component.Type == common.DevfileComponentTypeDockerimage
+}
 
-	return true
+// GetBootstrapperImage returns the odo-init bootstrapper image
+func GetBootstrapperImage() string {
+	if env, ok := os.LookupEnv(bootstrapperImageEnvName); ok {
+		return env
+	}
+	return defaultBootstrapperImage
 }
 
 // GetSupportedComponents iterates through the components in the devfile and returns a list of odo supported components
@@ -37,168 +68,4 @@ func GetSupportedComponents(data versions.DevfileData) (components []common.Devf
 		}
 	}
 	return components
-}
-
-// GetCommand iterates through the devfile commands and returns the associated devfile command
-func getCommand(data versions.DevfileData, commandName string, required bool) (supportedCommand common.DevfileCommand, err error) {
-	for _, command := range data.GetCommands() {
-		if command.Name == commandName {
-
-			// Get the supported actions
-			supportedCommandActions, err := getSupportedCommandActions(data, command)
-
-			// None of the actions are supported so the command cannot be run
-			if len(supportedCommandActions) == 0 {
-				return supportedCommand, errors.Wrapf(err, "The command \"%v\" was found but its actions are not supported", commandName)
-			} else if err != nil {
-				glog.Warning(errors.Wrapf(err, "The command \"%v\" was found but some of its actions are not supported", commandName))
-			}
-
-			// The command is supported, use it
-			supportedCommand.Name = command.Name
-			supportedCommand.Actions = supportedCommandActions
-			return supportedCommand, nil
-		}
-	}
-
-	// The command was not found
-	msg := fmt.Sprintf("The command \"%v\" was not found in the devfile", commandName)
-	if required {
-		// Not found and required, return an error
-		err = fmt.Errorf(msg)
-	} else {
-		// Not found and optional, so just log it
-		glog.V(3).Info(msg)
-	}
-
-	return
-}
-
-// getSupportedCommandActions returns the supported actions for a given command and any errors
-// If some actions are supported and others have errors both the supported actions and an aggregated error will be returned.
-func getSupportedCommandActions(data versions.DevfileData, command common.DevfileCommand) (supportedCommandActions []common.DevfileCommandAction, err error) {
-	glog.V(3).Infof("Validating actions for command: %v ", command.Name)
-
-	problemMsg := ""
-	for i, action := range command.Actions {
-		// Check if the command action is of type exec
-		err := validateAction(data, action)
-		if err == nil {
-			glog.V(3).Infof("Action %d maps to component %v", i, *action.Component)
-			supportedCommandActions = append(supportedCommandActions, action)
-		} else {
-			problemMsg += fmt.Sprintf("\nProblem with command \"%v\" action #%d: %v", command.Name, i+1, err)
-		}
-	}
-
-	if len(problemMsg) > 0 {
-		err = fmt.Errorf(problemMsg)
-	}
-
-	return
-}
-
-// validateAction validates the given action
-// 1. action has to be of type exec
-// 2. component should be present
-// 3. command should be present
-func validateAction(data versions.DevfileData, action common.DevfileCommandAction) (err error) {
-
-	// type must be exec
-	if *action.Type != common.DevfileCommandTypeExec {
-		return fmt.Errorf("Actions must be of type \"exec\"")
-	}
-
-	// component must be specified
-	if action.Component == nil || *action.Component == "" {
-		return fmt.Errorf("Actions must reference a component")
-	}
-
-	// must specify a command
-	if action.Command == nil || *action.Command == "" {
-		return fmt.Errorf("Actions must have a command")
-	}
-
-	// must map to a supported component
-	components := GetSupportedComponents(data)
-
-	isActionValid := false
-	for _, component := range components {
-		if *action.Component == *component.Alias && isComponentSupported(component) {
-			isActionValid = true
-		}
-	}
-	if !isActionValid {
-		return fmt.Errorf("The action does not map to a supported component")
-	}
-
-	return
-}
-
-// GetBuildCommand iterates through the components in the devfile and returns the build command
-func GetBuildCommand(data versions.DevfileData, devfileBuildCmd string) (buildCommand common.DevfileCommand, err error) {
-	if devfileBuildCmd != "" {
-		// a build command was specified so if it is not found then it is an error
-		buildCommand, err = getCommand(data, devfileBuildCmd, true)
-	} else {
-		// a build command was not specified so if it is not found then it is not an error
-		buildCommand, err = getCommand(data, string(DefaultDevfileBuildCommand), false)
-	}
-
-	return
-}
-
-// GetRunCommand iterates through the components in the devfile and returns the run command
-func GetRunCommand(data versions.DevfileData, devfileRunCmd string) (runCommand common.DevfileCommand, err error) {
-	if devfileRunCmd != "" {
-		runCommand, err = getCommand(data, devfileRunCmd, true)
-	} else {
-		runCommand, err = getCommand(data, string(DefaultDevfileRunCommand), true)
-	}
-
-	return
-}
-
-// ValidateAndGetPushDevfileCommands validates the build and the run command,
-// if provided through odo push or else checks the devfile for devBuild and devRun.
-// It returns the build and run commands if its validated successfully, error otherwise.
-func ValidateAndGetPushDevfileCommands(data versions.DevfileData, devfileBuildCmd, devfileRunCmd string) (pushDevfileCommands []common.DevfileCommand, err error) {
-	var emptyCommand common.DevfileCommand
-	isBuildCommandValid, isRunCommandValid := false, false
-
-	buildCommand, buildCmdErr := GetBuildCommand(data, devfileBuildCmd)
-
-	if devfileBuildCmd == "" && buildCmdErr != nil {
-		// If there was no build command specified through odo push and no default build command in the devfile, default validate to true since the build command is optional
-		isBuildCommandValid = true
-		glog.V(3).Infof("No build command was provided")
-	} else if buildCmdErr == nil {
-		isBuildCommandValid = true
-		// The build command is optional so there may not be an error but the build command may still be empty, only add it if it exists
-		if !reflect.DeepEqual(emptyCommand, buildCommand) {
-			pushDevfileCommands = append(pushDevfileCommands, buildCommand)
-			glog.V(3).Infof("Build command: %v", buildCommand.Name)
-		}
-	}
-
-	runCommand, runCmdErr := GetRunCommand(data, devfileRunCmd)
-	if runCmdErr == nil && !reflect.DeepEqual(emptyCommand, runCommand) {
-		pushDevfileCommands = append(pushDevfileCommands, runCommand)
-		isRunCommandValid = true
-		glog.V(3).Infof("Run command: %v", runCommand.Name)
-	}
-
-	// If either command had a problem, return an empty list of commands and an error
-	if !isBuildCommandValid || !isRunCommandValid {
-		commandErrors := ""
-		if buildCmdErr != nil {
-			commandErrors += buildCmdErr.Error()
-		}
-		if runCmdErr != nil {
-			commandErrors += runCmdErr.Error()
-		}
-		return []common.DevfileCommand{}, fmt.Errorf(commandErrors)
-	}
-
-	return pushDevfileCommands, nil
 }
