+++ pkg/devfile/adapters/kubernetes/utils/utils_test.go
@@ -1,0 +1,220 @@
+package utils
+
+import (
+	"reflect"
+	"testing"
+
+	"github.com/openshift/odo/pkg/devfile"
+	adaptersCommon "github.com/openshift/odo/pkg/devfile/adapters/common"
+	"github.com/openshift/odo/pkg/devfile/versions/common"
+	versionsCommon "github.com/openshift/odo/pkg/devfile/versions/common"
+	"github.com/openshift/odo/pkg/testingutil"
+
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestUpdateContainersWithSupervisord(t *testing.T) {
+
+	command := "ls -la"
+	component := "alias1"
+	image := "image1"
+	workDir := "/root"
+	validCommandType := common.DevfileCommandTypeExec
+	emptyString := ""
+	defaultCommand := []string{"tail"}
+	defaultArgs := []string{"-f", "/dev/null"}
+	supervisordCommand := []string{adaptersCommon.SupervisordBinaryPath}
+	supervisordArgs := []string{"-c", adaptersCommon.SupervisordConfFile}
+
+	tests := []struct {
+		name                    string
+		runCommand              string
+		containers              []corev1.Container
+		commandActions          []common.DevfileCommandAction
+		componentType           common.DevfileComponentType
+		isSupervisordEntrypoint bool
+		wantErr                 bool
+	}{
+		{
+			name:       "Case: Container With Command and Args",
+			runCommand: emptyString,
+			containers: []corev1.Container{
+				{
+					Name:            component,
+					Image:           image,
+					ImagePullPolicy: corev1.PullAlways,
+					Command:         defaultCommand,
+					Args:            defaultArgs,
+					Env:             []corev1.EnvVar{},
+				},
+			},
+			commandActions: []versionsCommon.DevfileCommandAction{
+				{
+					Command:   &command,
+					Component: &component,
+					Workdir:   &workDir,
+					Type:      &validCommandType,
+				},
+			},
+			componentType:           common.DevfileComponentTypeDockerimage,
+			isSupervisordEntrypoint: false,
+			wantErr:                 false,
+		},
+		{
+			name:       "Case: Container With Command and Args but Missing Work Dir",
+			runCommand: emptyString,
+			containers: []corev1.Container{
+				{
+					Name:            component,
+					Image:           image,
+					ImagePullPolicy: corev1.PullAlways,
+					Command:         defaultCommand,
+					Args:            defaultArgs,
+					Env:             []corev1.EnvVar{},
+				},
+			},
+			commandActions: []versionsCommon.DevfileCommandAction{
+				{
+					Command:   &command,
+					Component: &component,
+					Type:      &validCommandType,
+				},
+			},
+			componentType:           common.DevfileComponentTypeDockerimage,
+			isSupervisordEntrypoint: false,
+			wantErr:                 false,
+		},
+		{
+			name:       "Case: Container With No Command and Args ",
+			runCommand: emptyString,
+			containers: []corev1.Container{
+				{
+					Name:            component,
+					Image:           image,
+					ImagePullPolicy: corev1.PullAlways,
+					Env:             []corev1.EnvVar{},
+				},
+			},
+			commandActions: []versionsCommon.DevfileCommandAction{
+				{
+					Command:   &command,
+					Component: &component,
+					Workdir:   &workDir,
+					Type:      &validCommandType,
+				},
+			},
+			componentType:           common.DevfileComponentTypeDockerimage,
+			isSupervisordEntrypoint: true,
+			wantErr:                 false,
+		},
+		{
+			name:       "Case: Custom Command Container With No Command and Args ",
+			runCommand: "customcommand",
+			containers: []corev1.Container{
+				{
+					Name:            component,
+					Image:           image,
+					ImagePullPolicy: corev1.PullAlways,
+					Env:             []corev1.EnvVar{},
+				},
+			},
+			commandActions: []versionsCommon.DevfileCommandAction{
+				{
+					Command:   &command,
+					Component: &component,
+					Workdir:   &workDir,
+					Type:      &validCommandType,
+				},
+			},
+			componentType:           common.DevfileComponentTypeDockerimage,
+			isSupervisordEntrypoint: true,
+			wantErr:                 false,
+		},
+		{
+			name:       "Case: Wrong Custom Command Container",
+			runCommand: "customcommand123",
+			containers: []corev1.Container{
+				{
+					Name:            component,
+					Image:           image,
+					ImagePullPolicy: corev1.PullAlways,
+					Env:             []corev1.EnvVar{},
+				},
+			},
+			commandActions: []versionsCommon.DevfileCommandAction{
+				{
+					Command:   &command,
+					Component: &component,
+					Workdir:   &workDir,
+					Type:      &validCommandType,
+				},
+			},
+			componentType:           common.DevfileComponentTypeDockerimage,
+			isSupervisordEntrypoint: true,
+			wantErr:                 true,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			devObj := devfile.DevfileObj{
+				Data: testingutil.TestDevfileData{
+					ComponentType:  tt.componentType,
+					CommandActions: tt.commandActions,
+				},
+			}
+
+			containers, err := UpdateContainersWithSupervisord(devObj, tt.containers, tt.runCommand)
+
+			if !tt.wantErr && err != nil {
+				t.Errorf("TestUpdateContainersWithSupervisord unxpected error: %v", err)
+			} else if tt.wantErr && err != nil {
+				// return since we dont want to test anything further
+				return
+			}
+
+			// Check if the supervisord volume has been mounted
+			supervisordVolumeMountMatched := false
+			envRunMatched := false
+			envWorkDirMatched := false
+
+			if tt.commandActions[0].Workdir == nil {
+				// if workdir is not present, dont test for matching the env
+				envWorkDirMatched = true
+			}
+
+			for _, container := range containers {
+				if container.Name == component {
+					for _, volumeMount := range container.VolumeMounts {
+						if volumeMount.Name == adaptersCommon.SupervisordVolumeName && volumeMount.MountPath == adaptersCommon.SupervisordMountPath {
+							supervisordVolumeMountMatched = true
+						}
+					}
+
+					for _, envVar := range container.Env {
+						if envVar.Name == envOdoCommandRun && envVar.Value == *tt.commandActions[0].Command {
+							envRunMatched = true
+						}
+						if tt.commandActions[0].Workdir != nil && envVar.Name == envOdoCommandRunWorkingDir && envVar.Value == *tt.commandActions[0].Workdir {
+							envWorkDirMatched = true
+						}
+					}
+
+					if tt.isSupervisordEntrypoint && (!reflect.DeepEqual(container.Command, supervisordCommand) || !reflect.DeepEqual(container.Args, supervisordArgs)) {
+						t.Errorf("TestUpdateContainersWithSupervisord error: commands and args mismatched for container %v, expected command: %v actual command: %v, expected args: %v actual args: %v", component, supervisordCommand, container.Command, supervisordArgs, container.Args)
+					} else if !tt.isSupervisordEntrypoint && (!reflect.DeepEqual(container.Command, defaultCommand) || !reflect.DeepEqual(container.Args, defaultArgs)) {
+						t.Errorf("TestUpdateContainersWithSupervisord error: commands and args mismatched for container %v, expected command: %v actual command: %v, expected args: %v actual args: %v", component, defaultCommand, container.Command, defaultArgs, container.Args)
+
+					}
+				}
+			}
+
+			if !supervisordVolumeMountMatched {
+				t.Errorf("TestUpdateContainersWithSupervisord error: could not find supervisord volume mounts for container %v", component)
+			}
+			if !envRunMatched || !envWorkDirMatched {
+				t.Errorf("TestUpdateContainersWithSupervisord error: could not find env vars for supervisord in container %v, found command env: %v, found work dir env: %v", component, envRunMatched, envWorkDirMatched)
+			}
+		})
+	}
+
+}
