+++ pkg/kclient/supervisord.go
@@ -1,49 +1,10 @@
 package kclient
 
-import corev1 "k8s.io/api/core/v1"
+import (
+	"github.com/openshift/odo/pkg/devfile/adapters/common"
 
-const (
-	// The init container name for supervisord
-	supervisordInitContainerName = "copy-supervisord"
-
-	// The default image for odo init bootstrapper container
-	defaultBootstrapperImage = "jeevandroid/odo-init-image"
-
-	// Create a custom name and (hope) that users don't use the *exact* same name in their deployment (occlient.go)
-	supervisordVolumeName = "odo-supervisord-shared-data"
-
-	// The supervisord Mount Path for the container mounting the supervisord volume
-	supervisordMountPath = "/opt/odo/"
-
-	// The supervisord binary path inside the container volume mount
-	supervisordBinaryPath = "/opt/odo/bin/supervisord"
-
-	// The supervisord configuration file inside the container volume mount
-	supervisordConfFile = "/opt/odo/conf/devfile-supervisor.conf"
-
-	// The path to the odo init image contents
-	odoInitImageContents = "/opt/odo-init/."
+	corev1 "k8s.io/api/core/v1"
 )
-
-// GetSupervisordVolumeName returns the supervisord Volume Name
-func GetSupervisordVolumeName() string {
-	return supervisordVolumeName
-}
-
-// GetSupervisordMountPath returns the supervisord Volume Name
-func GetSupervisordMountPath() string {
-	return supervisordMountPath
-}
-
-// GetSupervisordBinaryPath returns the path to the supervisord binary
-func GetSupervisordBinaryPath() string {
-	return supervisordBinaryPath
-}
-
-// GetSupervisordConfFilePath returns the path to the supervisord conf file
-func GetSupervisordConfFilePath() string {
-	return supervisordConfFile
-}
 
 // AddBootstrapSupervisordInitContainer creates an init container that will copy over
 // supervisord to the application image during the start-up procress.
@@ -51,12 +12,12 @@ func AddBootstrapSupervisordInitContainer(podTemplateSpec *corev1.PodTemplateSpe
 
 	podTemplateSpec.Spec.InitContainers = append(podTemplateSpec.Spec.InitContainers,
 		corev1.Container{
-			Name:  supervisordInitContainerName,
-			Image: defaultBootstrapperImage,
+			Name:  common.SupervisordInitContainerName,
+			Image: common.GetBootstrapperImage(),
 			VolumeMounts: []corev1.VolumeMount{
 				{
-					Name:      GetSupervisordVolumeName(),
-					MountPath: supervisordMountPath,
+					Name:      common.SupervisordVolumeName,
+					MountPath: common.SupervisordMountPath,
 				},
 			},
 			Command: []string{
@@ -64,8 +25,8 @@ func AddBootstrapSupervisordInitContainer(podTemplateSpec *corev1.PodTemplateSpe
 			},
 			Args: []string{
 				"-r",
-				odoInitImageContents,
-				supervisordMountPath,
+				common.OdoInitImageContents,
+				common.SupervisordMountPath,
 			},
 		})
 }
