+++ pkg/occlient/occlient.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"io"
 	"net"
-	"os"
 	"path/filepath"
 	"reflect"
 	"sort"
@@ -22,6 +21,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/openshift/odo/pkg/config"
+	"github.com/openshift/odo/pkg/devfile/adapters/common"
 	"github.com/openshift/odo/pkg/log"
 	"github.com/openshift/odo/pkg/preference"
 	"github.com/openshift/odo/pkg/util"
@@ -106,13 +106,13 @@ const (
 
 	// Default Image that will be used containing the supervisord binary and assembly scripts
 	// use getBoostrapperImage() function instead of this variable
-	defaultBootstrapperImage = "registry.access.redhat.com/openshiftdo/odo-init-image-rhel7:1.0.2"
+	// defaultBootstrapperImage = "registry.access.redhat.com/openshiftdo/odo-init-image-rhel7:1.0.2"
 
 	// ENV variable to overwrite image used to bootstrap SupervisorD in S2I builder Image
-	bootstrapperImageEnvName = "ODO_BOOTSTRAPPER_IMAGE"
+	// bootstrapperImageEnvName = "ODO_BOOTSTRAPPER_IMAGE"
 
 	// Create a custom name and (hope) that users don't use the *exact* same name in their deployment
-	supervisordVolumeName = "odo-supervisord-shared-data"
+	// supervisordVolumeName = "odo-supervisord-shared-data"
 
 	// ComponentPortAnnotationName annotation is used on the secrets that are created for each exposed port of the component
 	ComponentPortAnnotationName = "component-port"
@@ -221,12 +221,12 @@ type Client struct {
 	Namespace            string
 }
 
-func getBootstrapperImage() string {
-	if env, ok := os.LookupEnv(bootstrapperImageEnvName); ok {
-		return env
-	}
-	return defaultBootstrapperImage
-}
+// func getBootstrapperImage() string {
+// 	if env, ok := os.LookupEnv(bootstrapperImageEnvName); ok {
+// 		return env
+// 	}
+// 	return defaultBootstrapperImage
+// }
 
 // New creates a new client
 func New() (*Client, error) {
@@ -1473,7 +1473,7 @@ func copyVolumesAndVolumeMounts(dc appsv1.DeploymentConfig, currentDC *appsv1.De
 			// Loop through all the volumes
 			for _, volume := range matchingContainer.VolumeMounts {
 				// If it's the supervisord volume, ignore it.
-				if volume.Name == supervisordVolumeName {
+				if volume.Name == common.SupervisordVolumeName {
 					continue
 				} else {
 					// check if we are appending the same volume mount again or not
@@ -1496,7 +1496,7 @@ func copyVolumesAndVolumeMounts(dc appsv1.DeploymentConfig, currentDC *appsv1.De
 
 	// Now the same with Volumes, again, ignoring the supervisord volume.
 	for _, volume := range currentDC.Spec.Template.Spec.Volumes {
-		if volume.Name == supervisordVolumeName {
+		if volume.Name == common.SupervisordVolumeName {
 			continue
 		} else {
 			// check if we are appending the same volume again or not
